pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "entities_api"
version = "1.28.1"
description = "Unified multi vendor AI inference API"
readme = "README.md"

authors = [
  { name = "Francis Neequaye Armah", email = "francis.neequaye@projectdavid.co.uk" }
]

license = { file = "LICENSE" }

keywords = ["AI", "API", "Assistants", "LLM", "RAG", "Vector"]

requires-python = ">=3.9"

dependencies = [
  "fastapi~=0.110.0",
  "SQLAlchemy",
  "pydantic>=2.10.3",
  "httpx>=0.24.1",
  "python-dotenv>=1.0",
  "aiofiles>=23.2.1",
  "qdrant-client~=1.13.3",
  "pdfplumber~=0.11.6",
  "validators~=0.34.0",
  "sentence-transformers~=4.0.1",
  "uvicorn[standard]",
  "ollama>=0.3.0",
  "groq",
  "pysmb",
  "typer>=0.12.0",
  "PyYAML>=6.0",
  "pymysql",
  "projectdavid-common",
]

[project.optional-dependencies]
dev = [
  "black==24.1.1",
  "isort==5.13.2",
  "ruff>=0.4.0",
  "pytest>=7.2",
  "pytest-cov",
  "mypy>=1.0",
  "bandit",
  "safety",
  "build",
  "twine",
  "types-PyYAML",
  "types-requests",
  "types-aiofiles",
]

[project.scripts]
platform-api = "entities_api.__main__:root"

[project.urls]
Homepage = "https://github.com/project-david-ai/platform"
Documentation = "https://github.com/project-david-ai/projectdavid_docs/tree/master/src/pages"
Changelog = "https://github.com/project-david-ai/platform/blob/master/CHANGELOG.md"

[tool.setuptools]
package-dir = { "" = "src/api" }

[tool.setuptools.packages.find]
where = ["src/api"]
exclude = ["tests", "docs"]

# ── Unified Formatting (SDK Parity) ───────────────────────────────────────
[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint.per-file-ignores]
"src/api/entities_api/daemons/*.py" = ["E402"]
"src/api/entities_api/cli/*.py"     = ["E402"]
"migrations/env.py"                 = ["E402"]

[tool.pytest.ini_options]
addopts = "--ignore=tests/integration"
pythonpath = ["src/api"]
